+++ src/admin/apis/AdminController.php
@@ -2,6 +2,7 @@
 
 namespace luya\cms\admin\apis;
 
+use luya\cms\base\BlockInterface;
 use Yii;
 use luya\cms\models\Layout;
 use luya\cms\models\Block;
@@ -85,6 +86,7 @@ class AdminController extends \luya\admin\base\RestController
                     continue;
                 }
                 // create the block object
+                /** @var BlockInterface $obj */
                 $obj = $block->getObject(0, 'admin');
 
                 // check if in hidden blocks
@@ -101,6 +103,7 @@ class AdminController extends \luya\admin\base\RestController
                     'id' => $block->id,
                     'name' => $obj->name(),
                     'icon' => $obj->icon(),
+                    'preview_enabled' => $obj->renderAdminPreview() ? true : false,
                     'full_name' => ($obj->icon() === null) ? $obj->name() : '<i class="material-icons">'.$obj->icon().'</i> <span>'.$obj->name().'</span>',
                     'favorized' => array_key_exists($block->id, $favs),
                     'newblock' => 1,