+++ src/admin/controllers/BlockController.php
@@ -3,6 +3,8 @@
 namespace luya\cms\admin\controllers;
 
 use luya\admin\ngrest\base\Controller;
+use luya\cms\models\Block;
+use yii\web\NotFoundHttpException;
 
 /**
  * Block Controller.
@@ -13,4 +15,25 @@ use luya\admin\ngrest\base\Controller;
 class BlockController extends Controller
 {
     public $modelClass = 'luya\cms\models\Block';
+
+    /**
+     * Generate the the block preview.
+     *
+     * @param int $blockId
+     * @return string
+     * @throws NotFoundHttpException
+     * @since 1.0.8
+     */
+    public function actionPreview($blockId)
+    {
+        $blockItem = Block::findOne($blockId);
+        if (!$blockItem) {
+            throw new NotFoundHttpException();
+        }
+
+        /** @var $blockObject \luya\cms\base\InternalBaseBlock */
+        $blockObject = $blockItem->getObject('preview', 'admin-preview');
+
+        return $blockObject->renderAdminPreview();
+    }
 }