+++ src/admin/resources/js/cmsadmin.js
@@ -1656,6 +1656,10 @@
 			$http.post('admin/api-cms-block/toggle-group', {group: group}, {ignoreLoadingBar: true});
 		};
 
+		$scope.isPreviewEnabled = function(item) {
+			return item.preview_enabled;
+		};
+
 		// controller logic
 
 		$scope.copyStack = ServiceBlockCopyStack.stack;