+++ tests/data/blocks/UnitTestBlock.php
@@ -18,6 +18,11 @@ abstract class UnitTestBlock extends InternalBaseBlock
     {
     }
 
+    public function renderAdminPreview()
+    {
+
+    }
+
     public function renderFrontend()
     {
     }