+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 
 + [#86](https://github.com/luyadev/luya-module-cms/issues/86) Add option to set a variation as default value `asDefault()`.
++ [#83](https://github.com/luyadev/luya-module-cms/issues/83) Show block preview on hover menu.
 
 ## 1.0.7.2 (17. October 2018)